refactor: merge and delete types and apply consistent usage across modules
- Deletes deprecated types: EncryptedImagePayload, EncryptedImage, wire payload splits
- Merges into unified types: EncryptedImage<T>, WirePayload, ImageDecryptionMeta
- Applies consistent usage across backend and frontend modules
- Cleans up unused files and renames others for clarity
- Updates all affected tests
