--- src/types/backend-wire-payload.ts
@@ -1,7 +1,0 @@
-import { ImageDecryptionMeta } from './image-decryption-meta.js'
-
-export interface BackendWirePayload {
-  encrypted: Buffer
-  meta: ImageDecryptionMeta
-  token: number
-}
