--- src/types/decrypted-image-payload.ts
@@ -1,6 +1,0 @@
-import { ImageFormat } from './image-format.js'
-
-export interface DecryptedImagePayload {
-  format: ImageFormat
-  buffer: Uint8Array
-}
