--- src/types/encrypted-image-payload.ts
@@ -1,6 +1,0 @@
-import type { ImageDecryptionMeta } from './image-decryption-meta.js'
-
-export interface EncryptedImagePayload {
-  encrypted: Uint8Array
-  meta: ImageDecryptionMeta
-}
