--- src/types/encrypted-image.ts
@@ -1,6 +1,0 @@
-import { ImageDecryptionMeta } from './image-decryption-meta.js'
-
-export interface EncryptedImage {
-  encrypted: Buffer
-  meta: ImageDecryptionMeta
-}
