--- src/types/image-encryption-meta.ts
@@ -1,5 +1,0 @@
-export interface ImageEncryptionMeta {
-  key: Buffer
-  iv: Buffer
-  tag: Buffer
-}
