+++ src/types/wire-payload.ts
@@ -1,5 +1,11 @@
 import { ImageDecryptionMeta } from './image-decryption-meta.js'
 
+export interface BackendWirePayload {
+  encrypted: Buffer
+  meta: ImageDecryptionMeta
+  token: number
+}
+
 export interface FrontendWirePayload {
   encrypted: Uint8Array
   meta: ImageDecryptionMeta
